+++ lib/usage_html.dart
@@ -29,20 +29,22 @@ class AnalyticsHtml extends AnalyticsImpl {
       new _PersistentProperties(applicationName),
       new _PostHandler(),
       applicationName: applicationName,
-      applicationVersion: applicationVersion);
+      applicationVersion: applicationVersion) {
+    int screenWidth = window.screen.width;
+    int screenHeight = window.screen.height;
+
+    setSessionValue('sr', '${screenWidth}x$screenHeight');
+    setSessionValue('sd', '${window.screen.pixelDepth}-bits');
+    setSessionValue('ul', window.navigator.language);
+  }
 }
 
 class _PostHandler extends PostHandler {
   Future sendPost(String url, Map<String, String> parameters) {
-    int screenWidth = window.screen.width;
-    int screenHeight = window.screen.height;
-    int clientWidth = document.documentElement.clientWidth;
-    int clientHeight = document.documentElement.clientHeight;
+    int viewportWidth = document.documentElement.clientWidth;
+    int viewportHeight = document.documentElement.clientHeight;
 
-    parameters['sr'] = '${screenWidth}x$screenHeight';
-    parameters['vp'] = '${clientWidth}x$clientHeight';
-    parameters['sd'] = '${window.screen.pixelDepth}-bits';
-    parameters['ul'] = window.navigator.language;
+    parameters['vp'] = '${viewportWidth}x$viewportHeight';
 
     String data = postEncode(parameters);
     return HttpRequest.request(url, method: 'POST', sendData: data);